slash.xyz/requirements.txt
python-dotenv
asyncio
pytz